requirements.txt
astroid==2.2.5
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.0
idna==2.8
isort==4.3.16
kaggle==1.5.1.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.3
oauthlib==3.0.0
pandas==0.24.2
pyasn1==0.4.5
pyasn1-modules==0.2.3
pylint==2.3.1
python-dateutil==2.7.5
python-slugify==2.0.1
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
tqdm==4.29.1
typed-ast==1.3.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.23
wrapt==1.11.1
